2nd draft of PDF decorators
