+++ src/wms/src/main/java/org/geoserver/wms/map/PDFMapOutputFormat.java
@@ -8,7 +8,9 @@ package org.geoserver.wms.map;
 import java.io.IOException;
 import org.geoserver.platform.ServiceException;
 import org.geoserver.wms.MapProducerCapabilities;
+import org.geoserver.wms.WMS;
 import org.geoserver.wms.WMSMapContent;
+import org.geoserver.wms.decoration.MapDecorationLayout;
 
 /**
  * Handles a GetMap request that spects a map in PDF format.
@@ -19,7 +21,14 @@ import org.geoserver.wms.WMSMapContent;
  * @version $Id$
  */
 public class PDFMapOutputFormat extends AbstractMapOutputFormat {
+    public PDFMapOutputFormat() {
+        super(MIME_TYPE);
+    }
 
+    public PDFMapOutputFormat(WMS wms) {
+        super(MIME_TYPE);
+        this.wms = wms;
+    }
     /** the only MIME type this map producer supports */
     static final String MIME_TYPE = "application/pdf";
 
@@ -40,6 +49,8 @@ public class PDFMapOutputFormat extends AbstractMapOutputFormat {
 
     public static class PDFMap extends org.geoserver.wms.WebMap {
 
+        public MapDecorationLayout layout;
+
         public PDFMap(final WMSMapContent mapContent) {
             super(mapContent);
         }
@@ -49,14 +60,13 @@ public class PDFMapOutputFormat extends AbstractMapOutputFormat {
         }
     }
 
-    public PDFMapOutputFormat() {
-        super(MIME_TYPE);
-    }
-
     /** @see org.geoserver.wms.GetMapOutputFormat#produceMap(org.geoserver.wms.WMSMapContent) */
     public PDFMap produceMap(final WMSMapContent mapContent) throws ServiceException, IOException {
 
         PDFMap result = new PDFMap(mapContent);
+        // TODO - can we find out if we are tiled here?
+        MapDecorationLayout layout = findDecorationLayout(mapContent.getRequest(), false);
+        result.layout = layout;
         result.setContentDispositionHeader(mapContent, ".pdf");
         result.setMimeType(MIME_TYPE);
         return result;
